+++ carts/tunnel.lua
@@ -56,8 +56,10 @@ function _update()
   -- Collision: did the nose of the ship end up inside a wall we drew?
   if ship < 3 or ship > FLOOR - 3 then
     dead = 30
+    lose()
   elseif pget(30, flr(ship)) ~= 0 and pget(30, flr(ship)) ~= 4 then
     dead = 30
+    lose()
   end
 end